Refactor ChatPage into modular hooks
