+++ src/utils/chat/greetings.ts
@@ -1,0 +1,6 @@
+export const saudacaoDoDiaFromHour = (hour: number) => {
+  if (hour < 6) return 'Boa noite';
+  if (hour < 12) return 'Bom dia';
+  if (hour < 18) return 'Boa tarde';
+  return 'Boa noite';
+};
